pyproject.toml
[project]
name = "youtube-music-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.1",
    "streamlit>=1.52.2",
    "tqdm>=4.67.1",
]
